+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 
 1. [](#new)
     * Added `$grav['login']->login()` and `$grav['login']->logout()` functions with event hooks
+    * Added `$grav['login']->getRateLimiter($context)` function
     * Added events `onUserLoginAuthenticate`, `onUserLoginAuthorize`, `onUserLoginFailure`, `onUserLogin`, `onUserLogout`
 1. [](#improved)
     * Remember entered username if login fails
+    * Improved rate limiter to work without sessions and against distributed attacks
 
 # v2.4.3
 ## mm/dd/2017